Issue #14 :: Showed in forms [BUGFIX]
- Adds changes to use the attribute to render the form element
- Adds changes to manage the enabled and all providers as select options
