+++ view/frontend/templates/customer_account/two_factor_auth_config.phtml
@@ -9,11 +9,12 @@
 /** @var \Magento\Framework\View\Element\Template $block */
 /** @var \Magetarian\CustomerTwoFactorAuth\ViewModel\Customer\Configuration $viewModel */
 $viewModel = $block->getViewModel();
+$providers = $viewModel->getProviderConfigAttribute();
 ?>
 <?php if ($viewModel->isEnabled()):?>
     <div class="block block-two-factor-auth">
         <div class="block-title">
-            <strong><?= $block->escapeHtml(__('Two Factor Authentication Providers')) ?></strong>
+            <strong><?= $block->escapeHtml($providers->getFrontendLabel()) ?></strong>
         </div>
         <div class="block-content">
             <div class="box box-providers">
@@ -25,11 +26,11 @@ $viewModel = $block->getViewModel();
                     <div class="box-content">
                         <?php $selected = $viewModel->getSelectedProviders(); ?>
                         <select name="providers[]" size="4"  multiple="multiple" id="providers-field" class="multiselect">
-                            <option value="" <?=(count($selected)==0 ? 'selected' : '') ?> ><?= $block->escapeHtml(__('None')) ?></option>
-                            <?php foreach ($viewModel->getAvailableProviders() as $providerCode => $providerName): ?>
-                                <option value="<?=$block->escapeHtmlAttr($providerCode);?>"
-                                    <?=(in_array($providerCode, $selected) ? 'selected' : '') ?>>
-                                    <?= $block->escapeHtml($providerName) ?>
+                            <option value="" <?= (count($selected) == 0 ? 'selected' : '') ?>><?= $block->escapeHtml(__('None')) ?></option>
+                            <?php foreach ($providers->getOptions() as $provider): ?>
+                                <option value="<?=$block->escapeHtmlAttr($provider->getValue());?>"
+                                    <?=(in_array($provider->getValue(), $selected) ? 'selected' : '') ?>>
+                                    <?= $block->escapeHtml($provider->getLabel()) ?>
                                 </option>
                             <?php endforeach; ?>
                         </select>
